list collections even if they contain no documents
Now that the journalist can delete documents, it's confusing if they delete all
documents. Then the source's collection "disappears" from the main list,
leaving them with no way to reply to the source unless the source sends another
submission.

it would be better to display the number of documents in the main list, and
allow journalists to sort the list of collections by the number of documents
contained.