Major update to install.md
We have decided to recommend the journalist log-in to the document interface using Tails. This means changing the Torrc file gets a lot more complicated because Tails securely wipes every time. So we made directions to create a file that journalists can save in their persistence and double click to give themselves access every time they log-in to Tails.