cleanup from removing util
* keeps datetime formatting in the template (w/ a custom template filter)
* fixes import error for @dolanjs
* should also fix Travis CI

So, the reason I couldn't reproduce the issues on Travis was because after
I git rm'ed util.py in a prior commit, util.pyc was left behind, which caused
everything to work (incorrectly) for me. Note that this manifested as
a "tests.test_unit not a module" error - looks like unittest swallows the
actual issue.