Update "you can encrypt with our public key" bottom msg.
The only people who should really do that are users already familiar w/PGP (as noted in the "why encrypt with key" page that we link to here). So let's make that explicit right off the bat so we don't scare people. Also mention that submissions become encrypted upon receipt.