Better names for dirs in config
Previously, we used `SECUREDROP_ROOT` for the directory where all of the
data (db, submissions, keyring) was stored. This caused confusion, as it
sounds like it refers to the root of the web application
(`/vagrant/securedrop` or `/var/www/securedrop`, depending on dev or
prod).

Changed `CODE_ROOT` => `SECUREDROP_ROOT`
Changed `SECUREDROP_ROOT` => `SECUREDROP_DATA_ROOT`