Loop until something is entered for the HOTP secret
If you choose "y" for "use HOTP", but then accidentally hit enter
instead of entering something for the HOTP secret, the program would
interpret the empty string as a "false-y" otp_secret value and revert to
generating a TOTP secret and displaying the QR code. This change loops
until something is entered for the HOTP secret to avoid this confusing
situation.