Do setup to run the functional tests on staging
This is not ideal, since the functional tests require dependencies that
will now be present on the staging server but would not be present on a
production server. For now, we'd like to be able to run all of the tests
to generate the best possible AppArmor rules, but we should revisit
this. Ideally, you would run the tests remotely, which should be fairly
easy to do with the functional tests at least.