Use separate job queue for jobs from manage.py test
Python-RQ is nice because it lets us schedule jobs that are arbitrary
Python functions. The downside is that, due to our messy "config as a
module with values derived from an environment variable" setup, there
can be issues. For example, jobs sent from the tests will not run
correctly on the default worker (started by supervisor), because they
were started with different values for SECUREDROP_ENV and therefore have
different SECUREDROP_ROOTs.

Clearly this whole situation needs to be improved but this hack is
reasonable and works well for now:

1. Use a separate "test" queue for jobs triggered by the tests
2. Start/stop a temporary "test"-queue-only rqworker for the tests