During one of the last merges I broke the procmail and the ossec sasl_password configuration. This fixes that.
Removed the .procmail file from being packaged with the ossec-server deb package. This was causing ansible not to reload postaliases.

Migrated the 'postalias /etc/aliases' and 'postmap /etc/postfix/sasl_passwd' task to a handler.

Changed the way that ossec alert gpg keyring is created so that it doesn't cause issues on subsequent runs and removed the warning comments about it.

Removed the unneeded backslash in the send_encrypted_alarm.sh.

Added creating the .gnupg directory in the OSSEC server's deb package postinst to avoid permission issues.