Save bulk downloads to SecureDrop-specific tmp dir
In order to serve bulk downloads with X-Send-File, we need to save them
in a SecureDrop-specific temp directory to avoid allowing Apache to
serve all of a default temp directory such as `/tmp`.

This commit also adds timestamps to the bulk download .zip filenames to
make them easier to work with.

Tested successfully on staging.