Add manage.py command to clean out SecureDrop's temp dir
This command is designed to be run as part of a regular cron job to
prevent this directory from accumulating past downloads and eventually
exhausting disk space. This command checks to see if a file is in use by
a process before deleting it, so it will not delete files that are in
the process of being downloaded.

Tested that this command cleans the tmp dir and does not interrupt
ongoing downloads on staging.