Comment out FPF repo setup in deb builder script
I did this for two reasons:

1. We aren't using any packages from the repo to build the debs at the
moment (if you think about it, that's kind of a catch-22), so commenting
this has no impact on building debs. Also, running `apt-get update`,
which is relativley slow, for no reason is a waste of time.

2. During testing, this created a catch-22 where I wanted to build the
debs so I could add them to the repo and publish the snapshot, but this
deb builder script wouldn't run unless the repo already existed.

We may be able to remove this code entirely, but I decided to leave it
in for now in case it is useful for something that I didn't recognize.