securedrop-grsec metapackage version should match grsec version
Previously the `Version:` of the securedrop-grsec kernel metapackage was
tracking the SecureDrop/securedrop-app-code version. Since the
metapackage tracks the grsecurity upgrades, and is independent of the
version of SecureDrop, it makes sense for the version of the metapackage
to match the grsecurity version.

Updated the deb builder (which we currently only use to build
securedrop-grsec) to get the version for the filename from the Debian
control file. This is important, since in my testing aptly refused to
add two packages with the same filename to the same repo (even if their
`Version:`s were different).

Also makes `build_meta` slightly more generic, to fit its name and
intended purpose. Also changed some variable names for clarity.