Single-use TOTP tokens
This is a hardening measure adapted from the google-authenticator PAM
module. By only allowing TOTP tokens to be used once, we can prevent or
detect MITM attacks.

There is a usability downside, as if you mistype your password you now
have to wait for the next TOTP token before attempting to log in again.
This is clearly explained in the flashed notification for TOTP users.

Improves the logging for journalist logins.