Throttle logins
Throttles login attempts on the document interface. Keeps a record of
each user's history of login attempts, and will throttle their attempts
if they make too many in a given time period (currently, more than 5
attempts in the last minute).

Also adds a flag for *all* login hardening so we can turn it off in the
tests, since the login hardening affects the tests unpredictiably and we
don't have time to rewrite the tests right now.