Reboot every 24 hours to periodically wipe memory
This is a kludge, but a reasonably effective one. This change reboots
the servers every day at 5am. Previously this reboot would only happen
if required by the daily package upgrade (e.g. for a kernel upgrade),
but now it happens every day regardless of whether it is needed.

This ensures that the plaintext of submissions can reside in memory for
at most 24 hours after submission. Resolves #99, although we can and
should do better in a future release.