removed ca_cert var.
We currently enforce checking the smtp relay's fingerprint in the postfix config

`securedrop/install_files/ansible-base/roles/ossec_server/templates/main.cf`

`smtp_tls_security_level = fingerprint`

One of the other options is to "verify" where it checks the certs ca chain. If verify is set the smtp_tls_CAfile option is required.
Since we set smtp_tls_Security_level to fingerprint the chain of trust is not checked. If smtp_cafile option is specified in the config file then a file does need to be provided but it is not used providing no value.