added a restore script
The 0.3_restore.py will restore the decrypted zip file created by the 0.3_collect.py script.

This script is different than the 0.3_migrate.py which migrated the different db layout and dir structure for a 0.2.1 install to a 0.3 version

This scripts restores the:
 securedrop config.py
 /var/lib/securedrop/store
 /var/lib/securedrop/keys
 /var/www/securedrop/static/i/logo.png
 torrc
 source interface tor certs
 document interface tor certs

This script does not restore the ssh hidden service from the backup. Deleting the previous ssh ATHS and replacing it from with the one from the backup would break the ssh session that ansible was using to run the playbook.