Update network firewall rules and docs
I updated the firewall rules slightly to match a working configuration
that was tested today. There are several options for how to set up the
rules to allow traffic from the Admin Workstation, and I was not sure
how to best express them in the iptables format, so I just added a
comment to the rules template. I also removed several lines of comments
that are irrelevant after recent changes.

I reworked the Network Firewall documentation completely. The primary
change was in recommending that users *not* use the default subnet of
192.168.1.1 for the LAN subnet, since it conflicts when connected to
another router that uses that (very common) subnet. I also explained
what is necessary to successfully configure the WAN in a both more
detailed and more abstract way. I added much more detailed documentation
for the initial configuration of the firewall. I edited the tips for
setting up the firewall rules based on some recent changes.

Finally, I added some screenshots from a working pfSense firewall
configuration.