Limit notifications for initial submission
We added a "first submission" reminder a while back as a friendly way to
remind sources that they should remember to return to their account and
check back for replies after making a submission. At the time, this
notification was shown in addition to the standard message and/or file
submission notifications. However, this resulted in a cluttered UI after
the first submission. After some experimentation, I decided it feels
natural to thank them for submitting something after the first
submission, and to use the standard message and/or file submission
notifications after that.

This commit changes the initial notification behavior on the Source
Interface and updates the tests accordingly.