Store .debs and build artifacts in new build directory
This fixes two annoying problems with the way building packages used to
work:

1. The built .debs and files downloaded as part of the build process
(e.g. the OSSEC .tar.gz) used to be stored in the top level directory of the
repo, which was a lot of clutter. Now they are stored in their own
dedicated directory, `build`.
2. In order to be used in conjunction with the `install_local_pkgs`
Ansible tag, the .debs used to have to be manually moved to
`install_files/ansible-base`, which was an annoying step that was easy
to forget. Now it looks in the build directory, which is where the debs
are put automatically.