apparmor updates
changed default mode for apparmor tor and apache2 profiles to enforce, but the ansible staging playbook will put the profiles in complain mode unless the skip tag `aa-complain` is skipped.

added the needed entries for the document interface and other parts of the web application to function.

added ossec decoders and rules to reduce alerts being sent for false positives.