Increase CSRF token time limit to 24h
The default timeout is 1h. Large uploads can take > 1h over Tor. It
would be incredibly frustrating to wait 1h+ for a large upload to
complete over Tor, only to have the upload rejected because the CSRF
token had expired while the upload was in progress.

24h was chosen arbitrarily, and may need to be tuned in the future based
on further testing of large uploads over Tor.