Improve submission deletion flow
Since submission deletion is a destructive action, we want to confirm it before
carrying it out. The original flow was cumbersome and required three
separate steps for each set of documents deleted:

1. Click delete
2. Confirm that the files you selected are the ones you want to delete
   on `delete.html`
3. See a confirmation on a separate page (!!) with a link that, when
   clicked, takes you back to the collection you started with.

This commit adds a new Javascript-based flow that uses a pop-up alert to
confirm deletion. The original flow, which did not use any
Javascript, has been left in in case the journalist's browser has JS
disabled, and has been improved by replace the 3rd step in the list
above with a redirect back to the original collection accompanied by a
flashed notification.

Choosing between these flows is handled automatically in `journalist.js`
(progressive enhancement).