Prevent apparmor errors on `/var/www/.gnupg`
It appears this issue was not completely fixed by
439a05bf4bc8d0d3b28bc5bb95a39d024b4b6609, as I had hoped, although the
incidence appears to have been reduced. We do not have time to
investigate the cause of this error now, so I am whitelisting the
directories that gpg-agent creates for now. Otherwise, this creates a
lot of logging noise (every time there is a server restart or a
submission).