moved tasks to ansible
moved tasks related to deleting the old singing key and the old package names to ansible tasks. This has the advantage that ansible will handle most edge cases gracefully.

The next time the playbook is run it will install the correct packages. This will help for sistuations in staging where someone is not running grsec.