custom header image example
This pr changes the handling of the branded custom header image to be treated like the securedrop-app-code config.py file

Some orgs use a branded custom header image for their source interface to distinguish their instance from other SecureDrop instances

Previously if a custom header image `securedrop_header_image` var was defined in the ansible yml config file then the playbook would overwrite the logo.png that was included in the `securedrop-app-code` deb package.

Since the logo.png was NOT set as a conffile in the debian package when automatice updates would happen the logo would be replaced.

Now only a logo.png.example image is included in the `securedrop-app-code` deb package. In the `initialize_securedrop_app` role if the variable securedrop_custom_header is set to `None` then ansible will copy logo.png.example to logo.png

This has the downside of the deb package not being self contained and requiring configuration from ansible for it to work, but it also means we don't need to worry about any custom header images being overwritten during updates.