add ssl support to apache apparmor profile
Some orgs are already using HTTPS for the source interface. This adds the required modules and default cert locations for apache to use ssl.

This pr does not add ssl support to the ansible playbook. Just the apparmor support is being included now so that when unattended security updates for securedrop-app-code.deb package are applied it does not break the existing installs using ssl.