Split css into separate files
Previously, the Source Interface and the Document Interface both used
the same CSS file (`/static/i/styles.css`). This sometimes created
problems where changing styles on one interface would cause problems with
the styles on another.

I've also updated the links in the HTML templates for each interface to
use the correct CSS file, depending on the interface.

In the future, we should work to clean up and simplify these CSS files.
I have not made an effort to remove code from `source.css` that was only
ever used on the Document Interface, nor vice versa.