Fix unmatched `*` in Markdown file
This document refers to "your specific .onion address" as
`*.onion`. Unfortunately, `*` is a reserved Markdown character that
signals that start of an italicized block. This can cause problems
depending on the Markdown renderer - for example, the document renders
to HTML fine when viewed on Github, but when this document is displayed
in a Github PR, all document text following that `*` is italicized,
which is confusing and not the intent.

The easy fix is to fence `*.onion` with backticks so it's treated as
preformatted text/code.