Synchronize upgrade docs in repo with draft in Hackpad
Upgrade docs draft:
https://securedrop.hackpad.com/Upgrade-to-0.3.3-jis3fmQ3AnE

I added a new `upgrade.md` to serve as a central place for all upgrade
information. It explains how the upgrades work at a high level
(incorporating text that was previously unused from the Hackpad draft)
before linking to the specific docs for upgrading to a specific version,
depending on the version you are upgrading from.