skip 'backup' role during upgrade playbook run
The rationale here is that if the Document Interface is not
regularly purged, then a complete backup over Tor would
cause the upgrade to fail. Updated docs and and comments
to upgrade script to explain this, and provide an example
command to run the backup manually if necessary.

The backup role is still enabled by default in the Ansible playbook.