Avoid breaking Tor when upgrading tails_files
Due to a race condition with the original, `.xsessionrc`-based approach
for auto-configuring Tor to use the SecureDrop authenticated Tor Hidden
Services, upgrading to 0.3.3 and re-running `tails_files/install.sh`
often caused Tor to break. This was confusing and meant that a restart
was required for an admin to restore a functioning Tor. For a full
explanation, see #1015.

This commit "repairs" the torrc backup used by `securedrop_init` if the
race condition occurred. This is a stopgap and only happens once, in the
session where we remove the broken .xsesssionrc while upgrading from
0.3.2 to 0.3.3's tails_files. We can safely remove this code after all
SecureDrop's are running SecureDrop >= 0.3.3.

This commit also checks if `.xsessionrc` is present before deleting
it. Previously, it would just delete the file and continue whether or
not there was an error (which there would be if the file was not
present). This works, but is poor shell scripting practice (not
"defensive coding"). Ideally `tails_files/install.sh` would have `set
-e` at the top to prevent it from continuing if an error occurs, but
that will have to be a long-term project due to the lack of error
handling throughout the rest of the script.