Use securedrop_app_code_version for OSSEC packages
The host-specific OSSEC configuration is provided in deb packages that
are versioned in lockstep with securedrop-app-code. These version values
were previously hardcoded in `host_vars/{app,mon}.yml`, but it is more
DRY to derive them from `securedrop_app_code_version`.

I moved the defintion of `securedrop_app_code_version` to
`group_vars/securedrop.yml` so it would be available when provisioning
both the app and monitor servers, and used it as a variable to generate
the OSSEC package names in `host_vars/{app,mon}.yml`.