Make 0.3pre upgrade script idempotent for logo.png
This is a hack to fix a specific issue that we encountered during
testing. The sequence of events is as such:

1. Run the upgrade playbook.
2. You potentially encounter #1008, which causes the playbook run to
fail after a reboot and leaves the upgrade in an unfinished state.

Ideally, you would be able to simply re-run the playbook to try the
upgrade again (and hopefully complete it). However, this is not possible
because:

1. The upgrade playbook removed the old SecureDrop .deb packages.
2. Due to #911, "logo.png" was included in the 0.3pre
securedrop-app-code .deb package, and so was removed when the packages
were removed.
3. Re-running `0.3pre_upgrade.py` fails when it tries to back up
`logo.png` in `backup_app`.

Since the custom logo.png file is also present on the Admin Workstation,
Ansible will place that file and therefore does not really need the copy
on the application server. Therefore, we can safely ignore this file if
it is not present. This should allow the script to conclude successfully
even if the upgrade had previously terminated early due to #1008.