remove paxtest install from app-test role
Since paxtest is used to validate the grsecurity config,
it's best placed in the grsecurity role. Whether it should
be included only in staging or in prod as well will be
discussed in a separate issue before reincluding.

This commit partially reverts 247b9975e89b98a0a18b90b74eece5d28b619172