Remove debugging statement for scrypt.hash error
This print statement was adding during the development of the journalist
login feature in an attempt to understand rare intermittent failures of
`scrypt.hash`. Since I had nothing to go on at the time, it prints all
available information from the function scope, which includes the
plaintext password. Although the error was very rare, this is not a good
practice and should not have been kept in production code.

Resolves iSEC-15FTC-2.