restrict the stagings env allow_direct_access rules
Added the default interface to the allow direct access rules for SSH, Source Interface and Document interface rules. This makes checking the counter for prod rule for that access over tor easier.

I did not add the interface to the DNS rules since it was not needed at this time. But adding the correct interface to all the rules is something we should do in another pr