updated OSSEC versions
A new version of OSSEC came out (2.8.2). Updated the OSSEC package names in the securedrop specific deb package control files. And in the ansible host var config files.

Tested to ensure that the deb package are created correctly and work in the staging environment.

The new version of OSSEC is 2.8.2+0.3.4