ensures prod_clean stage starts with direct ssh connections
The env var SECUREDROP_SSH_OVER_TOR is exported later in the
script to instruct Vagrant to connect to the ATHS values when
opening an SSH connection. At the start of the script, since
the /prod/ instances are destroyed, the environment should be
reset to prohibit SSH connections over Tor.