Redirect /lookup if visiting /generate when logged in
When a source is already logged in, clicking the back button could log
them out and generate a new codename. This could cause the source to lose
their original codename if it was not memorized yet. This commit will
redirect the source to the look page if they accidentally go back to the
generate page when logged in. It also adds an explicit logout button when
the source is logged in. This adds both confidence and professionalism to
the appearance.