adds documentation of `fingerprint` security level
The Postfix config defaults to `secure`, which should work
well out of the box for common configs. In case admins want to
restrict verification to a specific fingerprint (the previous
default behavior for the SecureDrop's OSSEC Postfix config),
these docs provide steps for how to do so.

Mostly reverts 153d17c, which removed the commands for querying
the fingerprint.