Chains cron-apt and reboot tasks in nightly cronjob
The previous config tried to implement cron-apt running
at 4AM every night, followed by a reboot at 5AM every night.
Chaining the two tasks isn't quite so simple, however, since
cron-apt automatically sleeps for up to an hour when run
noninteractively. (This sneaky behavior is easy to overlook
during testing, since running `cron-apt` from the command line
skips the sleep logic.)

The new config moves the reboot command to the same line
as cron-apt, chaining it with && to prevent reboots in the
event of a failed upgrade (which could leave the machine
in an unbootable state).