Re-add signal rule to Apache AppArmor profile
When Apache shuts down, it needs to be able to send a SIGTERM to its
worker child processes, which are currently unconfined by
AppArmor. Re-adding this rule allows Apache to work with AppArmor on
systems using an Ubuntu kernel, which has kernel support for the signal
rules. On other kernels (generic, grsecurty, etc.) this rule is ignored.

Fixes #1078.