prunes old comments about digitalocean plugin
The comments referred to an old version of the
`vagrant-digitalocean` plugin that supported
snapshots from DigitalOcean. Using snapshots
would speed up provisioning when testing upgrades
(but not when testing clean installs), but
that's a goal for down the road. Therefore we
don't need to clutter up live files with
outdated comments.