Uses group_vars for development and build hosts
Adds 'development' group to group_vars and Vagrantfile.
The 'development' and 'build' VMs share vars anyway,
so it makes sense to collapse them for easier reuse.

In the app-test role, rather than skipping the "non-development"
tag by default on the development host, only include the skipped
tasks if the current host is in the "staging" group.