Makes haveged config changes idempotent in app-test role
Previously the `lineinfile` task adjusting the DAEMON_ARGS
var in /etc/defaults/haveged appended the line every time,
causing the playbooks to always report "changed" status for
this task, and therefore always fire the handler for restarting
the haveged service. The duplicate lines don't seem to impair
the service, but it's poor practice.