Avoid potential problems with reaching the WebGUI
While following the network firewall guide, I found I was unable to
connect to the pfSense WebGUI after connecting the Admin Workstation to
the LAN. This was because I had already configured a wireless network
connection on the Admin Workstation, and it had the same private network
address as the LAN defaults. This meant that going to `192.168.1.1`
routed me to the wireless network connection's gateway, instead of the
LAN gateway.

The easiest way to avoid this problem is to make sure you're *only*
connected to the LAN before trying to reach the WebGUI, so this commit
adds a note explaining this.