Updates tests for Postfix security settings.
As of #1100, the Ansible config defaults to
trusting system CAs for verifying the SMTP relay,
rather than requiring a fingerprint for the TLS cert.
The spectests now check for the new settings.

The Ansible config still supports the fingerprint
verification setup, though, so the old checks are
left in the tests file commented out, to make
conditional tests easier down the road. I'd rather
have the tests fail once on a nonstandard config,
then be easily tweakable, then add lots of conditional
logic to the test suite that would itself require testing.