General improvements to `install.md`: intro and Tails disks
Lots of small changes and improvements to the main install document. Beefing
up the “before you begin” section to link to more of the other docs recently
created and to give a better sense of what's going on and what you're getting
into. Changes to the Tails disk setup instructions to more-closely resemble
actual practice and provide a clearer and more intuitive framework for
getting started.