Simplified outline for index TOCs
In the initial migration from markdown to ReST/Sphinx/RTD, the index
page just had a single TOC with every documentation file in one big
list. Now that we have Sphinx/RTD, we can take advantage of all of its
features to organize and interlink the docs in order to make them more
discoverable.

This outline is my first stab at doing this. I explain in an
introduction that this documentation is meant for end users
only (sources, journalists, and administrators), while developer
documentation will be maintained on the Github project wiki. I haven't
moved the "developer" docs to the Github wiki yet, but plan to do so in
future commits.

Once I'd established a focus on end user documentation, I decided to
split the index into two main sections:

1. A series of guides for each type of end user (Source, Journalist,
Administrator)

2. A big outline of all the documentation for the installation process

I also added placeholders for future documentation on troubleshooting
and a pre-installation overview.