Updates to the installation docs focusing on the SVS and Data Transfer Device
These changes standardize the use of terms like “Tails drive” unabbreviate
“Secure Viewing Station” and so on from the start of the install doc through
the end of the Data Transfer Device section.

This commit includes a new section in `install.md` explaining how to set up an
encrypted USB drive as a Data Transfer Device (including screenshots) and the
use of the persistent Gnome Keyring feature to avoid having to repeately enter
the LUKS passphrase.

References are also added to a (currently fictional) `onboarding.md` document.
That doc should explain how to set up journalists with their SecureDrop
accounts, hardware, and so on. More discussion is in issue #1135.