Removes Ruby/rake dependencies
The Rakefile implementation for looking up checksums
was helpful, but bundling a custom library for performing
the lookups automatically when the role is run is even better.

Updates README with new provisioning steps. The info
in the "Examples" section should be updated when the vars
are renamed throughout the role.