Use hkps SKS pool, remove redundant fingerprint verification
Tails ships with the SKS hkps pool certificate and is set to use the hkps pool
by default. Passing `--keyserver pool.sks-keyservers.net` opens the admin up to
MitM attacks, and so I have removed it.

Since August 2014, GPG's `--recv-key` will check that the fingerprint of the
received key matches the fingerprint passed on the commandline. Thus, it is no
longer necessary to separately run `gpg --fingerprint`.

This fixes the problems described in issue #520.

Signed-off-by: Noah Vesely <fowlslegs@riseup.net>