Add placeholders to redirect people from old docs
Before we migrated to ReadTheDocs in #1158, we regularly shared links to
the documentation as links to the Markdown files on GitHub (because
GitHub would render the Markdown for us automatically). In order to
assist people who may still be using those old links, I've added
placeholders for some of our more popular documentation (based on a
quick skim of my Sent email and the broken links cited in #1166) that
explain the migration and redirect users to the correct pages in the new
docs.