Bundle and minimize CSS, JS assets
Uses flask-assets to group the CSS & JS assets, run them through cssmin,
and link to them in the index.html and base.html templates for both
journalists and sources

Adds flask-assets, jsmin, and cssmin as securedrop depedencies

Flask-assets lazily compiles the assets when they're requested, and
writes them into a .webassets-cache directory in the static directory.
As such we extend the apparmor profile to allow rw permissions to the
cache directory, as well as the compiled JS and CSS assets for both the
source and document interfaces.