Consolidate tasks for sudoers config
The `copy` module supports the `validate` parameter, which will run a
custom command to assert a valid file prior to writing. If the check
fails, the copy task will abort before the file is written. Therefore
we don't need separate tasks to check the sudoers config file and write
it. Preserving the conservative approach via `backup: yes`.