Set logo sizes to prevent browser redraw on load
Currently the logo image served is twice the size of the element in
which it will render. This causes the browser to show it at full
resolution until the styleshet loads, at which point it redraws.

This change introduces width and height attributes to the logo img tags,
such that the image will always dispay at the same size