Idempotent fetching of hidden service files
Rather than fetch-then-format, which results in two changed tasks every
time the playbook runs, we can read the file contents, then write the
local file in precisely the format we want. The cat-then-write approach
ensures that the tasks are marked "changed" only if the file contents
have indeed been changed, which is what we want.