Idempotent service management for ossec-authd process
The OSSEC init scripts are structured such that Ansible's service module
can communicate with them. Running `service ossec status` shows a
variety of OSSEC-related services:

    ossec-monitord not running...
    ossec-logcollector not running...
    ossec-remoted not running...
    ossec-syscheckd not running...
    ossec-analysisd not running...
    ossec-maild not running...
    ossec-execd not running...

But nothing for ossec-authd. Therefore we'll use pgrep to check for
the authd process, and kill it if found. Requires two tasks rather than
one, but allows for idempotence in reporting changed status.