Idempotent management of paxctl flags
Rather than running the paxctl command every time, let's read
the flags first and run the command only if the flags are not correct.
There's an opportunity to be more explicit with the flags we're setting
here, since the -x and -e flags are present by default in VirtualBox,
but aren't explicitly set via the paxctl command.