Idempotent checking for OSSEC agent registration
The list_agents command is read-only, so we can ignore the changed
status it always reports. There was also a bit of a logic error in the
subsequent tasks, as they weren't actually checking the STDOUT of the
list_agents command. Fortunately there are two conditionals on most of
those tasks (creates=), so they were skipped anyway, but the logic is
now fixed.