Allow full test suite to be run locally
Previously the snap_ci scripts required several manual tweaks for
running offline. It's cleaner to detect the Snap-CI environment via the
"SNAP_CI" env var and automatically clean up droplets if detected. If
not detected, the droplets are retained, which makes repeated use of the
script for running all spectests much more convenient during
development, particularly of the Ansible playbooks.