Better formatting on table in OSSEC alerts guide
The previous table formatting was converted from Markdown, and had no
linewrapping for table entries. This change provides linewrapping, and
while not perfect, is much more readable.

Unfortunately the rendered text still results in a horizontal scrollbar
in modern browsers, but the width of the scrolled box is not as wide as
before. So, a definite improvement. Not going to mess with custom CSS
at this time.