Properly ensures absence of apache default web dir
The /var/www/html default directory gets created automatically when
apache is installed. Previously the `app` role removed the directory as
part of the app dir tasks, but apache is installed after that. This
commit moves the removal task to after the apache install task, and
adds a spectest to ensure it's set.