Updates spectests for apt repo presence
The install-local-packages role supersedes the install-fpf-repo role in
the Ansible config, so the spectests must also have it as a separate
role. It's skipped on the `*-staging` hosts in DigitalOcean.

Also updates the Snap-CI test script to skip the grsec role only on
DigitalOcean. The Rakefile for the spectests also inspects the provider
and drops the corresponding tests.