Removes vagrant up commands for build machine
Automatically bringing up the build machine sounds convenient, but
doesn't actually work as intended. If the bulid VM was not present
during a staging run, the `vagrant up` command will fix that, but it'll
still be missing from the automatically generated inventory file that
Vagrant creates, causing the `groups.all` check to fail anyway.

Therefore it's best to check and fail, but skip the automatic `up`.