Serverspec: adds regression tests for apparmor profiles in build role
The contents of the AppArmor profiles are tested in the staging
spectests, but it's useful to validate the presence of the profiles
during the build tests, strictly to guard against regressions. This is
particularly important when refactoring the build process itself.

Also adds a note about aa-complain tag in task comment.

Noticed this while testing: aa-complain happily exits 0 even if you pass
it a nonexistent path for the profile. Which is utter madness. Added a
TODO to come back and address later, as part of the idempotence cleanup.
Registering task results with with_items is a bit tricky, so didn't
shoehorn the change into this branch.