Updates rsync version on CentOS for Snap CI
Required for synchronizing the Debian package files via rsync. A
workaround would be to hard-code all filepaths, recursively create the
parent directories, then manually enforce the permissions after copying.
It's much simpler and more reliable to let rsync do the heavy lifting on
the file syncing.

So we'll pull down a recent version of rsync in the Staging environment
for the CI tests, but it's not necessary in the other stages.