Updates permission handling for procmail log file
The previous task was not correctly specifying file mode in
string-format, so "0660" was interpreted as "1224" on the host
(that's --w--w-r-T, to save you the trouble). Fortunately the wrong
permissions didn't break OSSEC, but it's definitely not ideal.

Now using a two-pass solution that doesn't clobber the logfile, but
ensures correct permission and ownership in string format.