Adds URL to developer docs to build playbook error message
First-time developers will likely not have built the OSSEC Debian
packages, which are managed in a separate repository. For local
development, the OSSEC Debian packages are required for the staging
environment. Rather than fail with an inscrutable file not found error,
detect the absence of the build machine from the play and explain the
situation. Include a link to the docs so new devs can get up to speed .