Serverspec: Enables regression tests for securedrop-app-code contents
The regression tests for validating the contents of the Debian package
for securedrop-app-code were previously commented out, and are now
enabled and passing. The contents were previously dumped from the 0.3.5
deb file, and stored in `vars/build.yml` in the spectests directory, in
order to guard against regressions.

This commit adds a few permitted files to the deb package contents,
files that were added after the release of 0.3.5. Explicitly:

  New pip requirements:
    Flask-Assets
    cssmin
    jsmin
    webassets (required by Flask-Assets)

  New coverage test:
    .coveragerc

  Sundry .pyc files
    securedrop/management/__init__.pyc
    securedrop/management/run.pyc

All regression tests are now passing.