Makes setting aa-complain in app-test role idempotent
Uses a messy shell command that pipes the output of `aa-status` to
`sed`, but it works. Another approach would be to compare the checksums
on the profile modes themselves, but that's not as safe, since it's
possible to update the profiles on disk but not reload the service, in
which case the states don't match. Safer to call aa-status and use its
output, even if it requires a janky one-liner.

Includes mad comments explaining the aa-status|sed shell command.
Definitely a contender for gnarliest one-liner of the year.

Also prunes tags on the task lists by consolidating the aa-complain tags
on the include in main.yml within the app-test role.