Fixes permissions on OSSEC email alert script
Previously the template task to write the script was setting 1407
permissions, rather than 0750. This is because the "mode:" parameter for
the task declared the value as an integer, rather than as a string.

We definitely do not want others to have write capability on this
script. Fixed, and updated the Serverspec test suite to match.

Using "0750" with ossec:root ownership. Typically "0755" would be
sufficient, but this script should be executed only as part of the OSSEC
functionality, and therefore should not be generally executable by any
user of the system.