Mask codename on login and submission pages
This commit masks sources' codenames altogether as they enter them into the
login page. On the submission page, it hides the codename by default, but allows
the source to reveal the codename via a hover-to-reveal functionality.
Hover-to-reveal was chosen over click-to-reveal with the intention of minimizing
overall visibility time.

It doesn't hurt they are now unable to copy-paste the codename from the
submission interface—although they still may do this during codename generation
process, this practice should be discouraged.

Signed-off-by: Noah Vesely <fowlslegs@riseup.net>