Moves iptables rules for Monitor Server into role
Use the consolidated template for iptables v4 rules to managing the SSH
direct access exemptions for the staging environment. Using a dual
conditional so that "staging" must be in the group names, and
"allow_direct_access" must be set to true. Both conditions default to
false, to avoid accidental misconfigurations.