Converts Travis CI tests to trusty64 with sudo
The Travis CI tests have been running on legacy infrastructure. Indeed,
each and every Travis run reminds you:

   This job ran on our legacy infrastructure. Please read our docs on
   how to upgrade. [0]

We've been dragging our feet on the upgrade, because the new
Docker-based containers explicitly do not allow the use of sudo, which
is a hard requirement for our tests. But sudo-less Docker containers are
not our only option (praise the old gods and the new)! We also have the
option of using Ubuntu 14.04 LTS builds with sudo enabled [1], which is
a much better option for our use case, given that we explicitly support
14.04 for production instances, and our development VMs also run 14.04.

[0] https://docs.travis-ci.com/user/migrating-from-legacy
[1] https://docs.travis-ci.com/user/trusty-ci-environment/