Handles authd cleanup via delegate_to in ossec-agent role
Rather than having a separate role targeting the Monitor Server
(remove-authd-exemptions) that runs on every invocation of the playbook,
we can inspect the state of OSSEC agent registration and clean up after
authd dynamically.

The iptables rules are now consolidated in the restrict-direct-access
role, so removing the authd roles prior to writing the iptables template
isn't necessary.

Logically we now have tasks that apply to the ossec-agent
and to the ossec-server in the ossec-agent role, so this commit also
breaks up the task list for that role into separate files.