Stick to Tails defaults for GPG keyserver options
In our last docs release, we switched to a hidden service SKS mirror. The
reasons for doing so were:

* Traffic never leaves the Tor network
* End-to-end TLS w/ self-authentication via the .onion URL

The hidden service keyserver has since gone down and this has caused a problem
for users trying to install SecureDrop, especially because GPG doesn't actually
inform the user when they specify an invalid keyserver address. Instead, they
are told that the operation they wanted to perform (e.g., `--search-keys`,
`--recv-key`) failed.

--

I propose we remove the `--keyserver` option altogether for the reasons that
follow.

In the Tails `gpg.conf` they default to using the SKS keyserver pool over
SSL/TLS (HKPS). The configuration file also specifies to replace the default
certificate store with just the singular self-signed certificate that is shared
among the TLS-capable SKS keyservers. This solves the CA problem in a different
way than the use of hidden services does. Traffic does leave the Tor network,
but only in encrypted form.

Further, the configuration specifies the `no-try-dns-srv` keyserver-option,
which though only documented in an upstream commit, prevents the leaking of DNS
(see https://trac.torproject.org/projects/tor/ticket/2846).

Lastly, sticking with the Tails `gpg.conf` defaults will be reliable since I
don't expect the entire SKS keyserver infrastructure to go down anytime soon.
Any problems our users have will be affecting all Tails and/ or SKS users, so we
can expect a timely fix in the unlikely chance of future problems.

--

One final thing to consider is the case that it has been discovered that the SKS
keyserver certificate is compromised. In this case, the certificate file will
not be replaced until the next version of Tails. It will be big news if this
does happen, so I'm sure we'll find out and quickly push a warning and
workaround to the documentation. Further, the risks implied by this problem are
still present with the use of hidden services, HKPS w/ the system certificate
store (as opposed to a self-signed cert file), and with no HKPS at all.

Signed-off-by: Noah Vesely <noah@freedom.press>