Uses template for formatting local THS files
The THS and ATHS config files fetched back to localhost should be
formatted with a trailing newline to support concatentation in the torrc
file. (The tails_files scripts open the files for copy/pasting, and
missing newlines will make it difficult to copy/paste correctly.)

Unfortunately Ansible v1 and v2 handle trailing newlines differently, so
the safest method to ensure reliable behavior across versions is to use
a template, and omit the hardcoded newline altogether. This approach
works on both v1 and v2.