Forces dest path for THS config files beside playbook
The behavior for inferring the dest path changed slightly in Ansible v2.
When running the playbook via vagrant, the *ths files will land adjacent
to the Vagrantfile, regardless of where the vagrant command was invoked
from. To force consistency between running ansible via the command line
and vagrant, force a path relative to the role directory and ensure the
files land adjacent to the playbook, as the docs indicate.

This change breaks portability for the role, but it's only used in
SecureDrop, so that's not currently a problem.