Replace old Markdown docs with RTD from develop
The last stable release, 0.3.6, was tagged before we rewrote
SecureDrop's documentation to use Sphinx and the ReadTheDocs platform
(#1158). The new docs are *much* better, and we currently point admins
to them instead of the Markdown docs available under the 0.3.6 tag in
the GitHub repository.

The problem: currently the only ReadTheDocs documentation available is
built from the develop branch, which is under active development. While
the installation and usage instructions are currently accurate for
0.3.x, they have already begun to diverge as develop moves toward 0.4,
and will only continue to diverge. ReadTheDocs has nice automatic
support for building "stable" documentation based on the latest tag
(according to semver semantics).

By copying the current ReadTheDocs documentation from develop into
0.3.7, we ensure that when 0.3.7 is released, we will have two sets of
documentation on ReadTheDocs: a fixed set for the stable release (based
on the docs that this commit backports from develop), and a "latest" set
based on develop, which is useful for reviewing docs changes.