Create backup_mon role, handle restore errors by reverting files
This commit builds on the 660f78d3a4e3e9e9a9991edad60ace5595340db8 and addresses
the issues raised in https://github.com/freedomofpress/securedrop/pull/1273.

**Maintaining SSH connections during Tor reloads**

In `/lib/systemd/system/tor@.service` we see the line:

`ExecReload=/bin/kill -HUP ${MAINPID}`

The default action on receiving a SIGHUP is to terminate the process, but
processes may trap the signal and handle it however they like. Daemons commonly
will reload their configuration files. In Tor’s `src/or/main.c` we see the
function, `do_hup` that handles SIGHUPs. In it the only subroutine that deals
with open circuits is:

```c
  /* Rotate away from the old dirty circuits. This has to be done
   * after we've read the new options, but before we start using
   * circuits for directory fetches. */
  circuit_mark_all_dirty_circs_as_unusable();
```

Note this does not close currently open circuits, it simply marks them such that
no new connections may be made over them:

```c
/** Mark <b>circ</b> so that no more connections can be attached to it. */
void
mark_circuit_unusable_for_new_conns(origin_circuit_t *circ)
{
  const or_options_t *options = get_options();
  tor_assert(circ);

  /* XXXX025 This is a kludge; we're only keeping it around in case there's
   * something that doesn't check unusable_for_new_conns, and to avoid
   * deeper refactoring of our expiration logic. */
  if (! circ->base_.timestamp_dirty)
    circ->base_.timestamp_dirty = approx_time();
  if (options->MaxCircuitDirtiness >= circ->base_.timestamp_dirty)
    circ->base_.timestamp_dirty = 1; /* prevent underflow */
  else
    circ->base_.timestamp_dirty -= options->MaxCircuitDirtiness;

  circ->unusable_for_new_conns = 1;
}
```

I think this is a conclusive explanation of why the SSH connection was
maintained for you @garretr after the reload, and sufficient to conclude that
this is the expected behavior we should be able to rely on.

**Keeping ATHSs for SSH accessible**

If an error happens during the restore process, we would like our admins to
retain the ability to remotely access the SD instance. This is implemented as
follows. Now, if a restore file is defined, the backup which is always done
before a restore, is not deleted from the remote server until the restore
process is over. This way, in case of an error, we can recover from this backup
file, reverting the files overwritten by the faulty restore_file. An error
message is also printed informing the admin that the restore was unsuccessful
and that we've reverted the files back for them.

**Keeping the servers in sync with the admin workstation**

If the restore process completed successfully, we re-use the code which
originally syncs the onion address and auth cookies to the
`./install_files/ansible-base/*ths` files during the install process (as well as
the tasks that prepend `HidServAuth` to the appropriate files).

**Sets group_var backup_script_version**

I think that setting a group_var makes more sense than sourcing `version.py`
because this means not needing to bump the version of the backup script w/ every
version bump.

**Replaces restore script with ansible tasks**

Since the functionality of the restore script could be reproduced with pure
Ansible, it made sense to get rid of it.

**Restore and backup for the Monitor Server as well**

It makes sense, especially in order to better support requests like multiple
admin support (see https://github.com/freedomofpress/securedrop/issues/1175), to
support backup and restore of Tor services and configuration for the mon server
as well.

**Adds support for testing backup and restore from staging VMs as well**

Because sometimes you don't want to test over Tor until you've tested locally
many times over. Soon you will be able to use the `SECUREDROP_SSH_OVER_TOR`
environmental variable as well to easily switch between accessing your staging
machines locally and via Tor, making this all the more handy.