Tails v2 upgrade review (#2)
* Adds the Tails v2 upgrade guide to the docs navbar

Places link to the Tails upgrade guide under the "Upgrade SecureDrop"
section. The title is automatically inferred from the ReST file, and
displays as "Upgrade Tails from 1.x to 2.x".

* Lighter language in introductory paragraph for Tails upgrade

Less strict terminology surrounding version numbers, and terser
sentences describing the steps to be taken. Lines exceed 80 chars, but I
don't want to reformat in a content-specific commit.

* More verbose list of required hardware for Tails upgrade

Explains the intended use of each component. In particular, the "master"
nomenclature for the Tails USB used to perform the upgrades via cloning
was not adequately explained. The brief in-line appositive now addresses
this concern, so Admins have a rough idea of the workflow before getting
started.

* Cull quotes from "master" Tails terminology

The "master" term is informal, which is good for our straight-shooting
admin docs, but the quoting was a bit excessive. Preserving quotes in
the initial occurrences, then removing thereafter.

* Move USB stick reuse advice into a Tip block

It's practically parenthetical advice, so it makes sense to separate it
from the main flow via Tip. Also added a small transition clause to what
is now the first sentence to set the scene for Admins: we're working in
the master Tails USB device from this point on.

* Uses boldface font for UI elements throughout Tails upgrade

Quoth the SecureDrop Documentation Guidelines:

  When specific elements from a user interface are mentioned by name or
  by label, **bold** it.

This commit enforces that pragma throughout the Tails 1.x -> 2.x guide.