0.3.7 reqs update, new pip-tools pkg mgmt system
Around `pip-tools` 1.0 they released some new utilities detailed in a blog post
http://nvie.com/posts/better-package-management/. Use of `pip-compile` allows us
specify only the packages we actually import in the SD app code, our top-level
dependencies, and not the lower-level dependencies (the dependencies of our
top-level dependencies).

Thus our top-level dependencies have been moved to `.in` files. Calling
`pip-compile` on a `.in` file will first find the latest versions of all the
packages declared in that file, and second find the latest compatible versions
of the lower-level dependencies, compiling them all together in a standard
`.txt` requirements file. This is beneficial because the dependencies of our
top-level packages may shift and this automates the process of finding and
tracking current and new lower-level dependencies and deprecating old
lower-level dependencies.

This commit also updates `pip_update.sh` to work with this new dependency
management system.