Creates function for determining Admin/Journalist Workstation
Previously the check for Admin versus Journalist Workstation was
incorrectly referencing the "app-document-aths" file on disk, which is
likely to exist in both environments, and therefore not a reliable
indicator. The check also exported a global var "ADMIN" that was
referenced in several different contexts.

Now, we're instead using a function to perform the check dynamically,
and the function returns 0 or 1, depending on the presence of
"app-document-aths". Further testing is needed to ensure that this is
still a reliable indicator. If not, we'll just force invocation of the
script to provide an explicit statement about the context.