Autoremove unused pip dependencies in app code postinst script
This addition to the `postinst` script in the app code package removes all pip
packages that neither belong to the OS (i.e., were installed by system packages)
nor are present in the requirements file in the package that was just unpacked.
The script itself just tries to uninstall all pip packages not declared in the
requirements file, but pip is smart and will not let you uninstall OS owned
packages.

Use of arrays required bash, which has been added to the `Depends:` just to be
thorough. And a duplicate copy of `python-pip` was removed from `Depends:` as
well.