Multibranch logic for Document Interface ATHS lookup
The Admin and Journalist workstations require slightly different setups
for ATHS values: the AW will have SSH info in /etc/tor/torrc, but
the JW will not; the AW will have flat files with ATHS info in
ansible-base; the JW will not; the AW will have Document Interface info
in flat files; the JW *may* have those files, but may not.

Accordingly we need to provide multiple attempts to lookup the existing
values if found, to avoid unnecessarily prompting Journalists for ATHS
info which they may or may not know how to look up themselves.