Strengthens logic for finding Document Interface on JW
The Journalist Workstation may already have the ATHS info for the
Document Interface in one of several locations:

  * /etc/tor/torrc
  * Desktop icon
  * app-document-aths flat file

Check all of them. If none are found, then we're probably running this
script for the first time, so go ahead and prompt for Journalist to type
in the information required. Originally this script was supposed to be
run once and only once, but certain changes have had to be rolled back,
thus requiring a strategic rewrite to make the script relatively
idempotent.

Uses an ad-hoc approach to memoization for the "prompt" function to
avoid reprompting. Stores the value in a global. Sue me.