Strengthens logic for finding Source Interface URL on JW
The Journalist Workstation may have the Source URL in a variety of
locations, whereas on the Admin Workstation we can much more confidently
assume that it'll be in `install_files/ansible-base/app-source-ths`.
Therefore we now have a multibranch approach to try to determine the
Source URL from several locations, finally falling back to prompting the
Journalist to enter in the URL manually if it cannot be found.

These changes additionally make this Bash script close to idempotent.
That was not previously a goal of the script, since it was intended to
be a "run-once" type of setup, but recent updates require that it be run
again to clean up mistakes in the past (e.g. ordering on NM hook).