Creates function for managing NetworkManger hook
The NM hook has undergone a name change to prevent a significant delay
in running to do an ordering mistake. The new name now causes the
SecureDrop hook to run AFTER Tor config but BEFORE apt update, which is
what we want. Included a verbose comment explaining the rationale.

The backwards-compatible legacy changes will need to be housed
in a separate function dedicated to cleanup.