Properly quotes all concatenated variables
In instances where variable names are concatenated with arbitrary
strings during interpolation, Bash best practices dictate enclosing the
variable name in brackets, and the entire concatenated string in double
quotes. This pragma is now enforced through the script.

There are still a few instances of unquoted vars, but they are not
concanenated, and also do not contain whitespace, since most are
constants declared at the top of the script.