update_version.sh autoinstalls deps, prepares changelog.md
Now instead of telling us to install devscripts and git, update_version.sh does
it itself. Since we're only running it from our Ubuntu build VM, we don't need
to worry about system compatibility.

This also puts the subheader for the new version in changelog.md and sets vim to
drop us right into the place in the changelog we're going to be writing in.

Because we are lazy, lazy devs.