Modify crypto_util.encrypt() to work with py-gnupg 2.0.2
We've been using an old branch of python-gnupg for quite some time for reasons
explained in
https://github.com/freedomofpress/securedrop/issues/250#issuecomment-86237456.
For over a year now https://github.com/freedomofpress/securedrop/issues/946 has
been open and just in 91f6491beb856d94f8bf6b38e162d1e23e5a581d I updated us to
the latest version.

Changes to GPG.encrypt() required changes in crypto_util. In particular,
GPG.encrypt() became the singular encryption function--there is no longer a
separate GPG.encrypt_file(). Further, GPG.encrypt() now handles encoding for us
using our preferred locale, which is already set on Ubuntu and any sane distro
to UTF-8--that's to say it's not something we should worry about.

Related to the ability for GPG.encrypt() to now be able to handle files and
other stream-like types, it needed to have some idea of what those types are.
This idea is implemented in the form of the list gnupg._util._STREAMLIKE_TYPES.
If a type is not in that list python-gnupg assumes it's a string-like type and
tries to encode it for you. Thus, it was necessary to
append(_TemporaryFileWrapper) to this list to avoid calling an undefined method
on SecureTemporaryFile (stream-like type membership testing is done with
isinstance() which holds true for subclasses as well).