Autoexit virtualenv on pip_update.sh invocation
Being in a virtualenv while running this script can causes errors. It now
detects if you are and deactivates it. Also, I cleaned up the tests for pip and
virtualenv install presence, so that it will try to install them automatically
on Debian and derivates, and will just tell the user to install them themselves
on other systems.