Expands comments in tails install script
No functional changes, simply trying to be more expressive in the
justification for this lengthy Bash script. With the new comments, each
section of the script should be intelligible to maintainers and new
contributors alike. Ideally I'd like to see this entire script replaced,
but until that fortuitous day, we'll use strong comments to guide our
understanding.