Tries harder to avoid prompting on Journalist Workstation
For existing Journalist Workstations running this script a second time
(necessary after upgrading to Tails v2), the HidServAuth info may
already be present in /etc/tor/torrc from a previous configuration.
Check there first and reuse that value if found. Otherwise, we'll prompt
for the info, which will also cover the case of new installs.

Strengthens the regex used for reading from the Source Interface desktop
icon. Before thes script assumed that the desktop icon was created properly.
Adds actual use of the global var for the Source URL, so we don't have
to prompt for it more than once.