Add backup tag to display_onions tasks
This allows an admin to run a backup and/or restore using the backup tag
alone. Adding the backup tag to the display_onions tasks ensures that
the `*ths` files are synchronized on the Admin Workstation after a
restore operation.

It is desirable to be able to run the backup/restore restore operations
in isolation because after installation is complete and the
restrict_direct_access roles are run, Ansible must run over SSH over
Tor, which means running the entire playbook can be very slow. It is
unnecessary to run the entire playbook to do a backup or restore, and it
would be frustrating to admins to have incur an unnecessary additional
time cost for every backup and restore, which are already often very
time-consuming due to transferring the potentially very large backup and
restore files over Tor.