Adds verbose output for A/THS info gathering
The logic for looking up Source and Document Interface information, such
as URLs and HidServAuth tokens, is rather convoluted. To aid in
debugging for Admins and reduce unintended surprises from automagical
configuration, have the script print out the locations from which the information
was extracted. This may be particularly relevant during credential
rotations, or upgrades.