Adds verbose output for SSH alias creation
The logic for configuring SSh aliases (in ~/.ssh/config) is cautious, to
avoid destroying custom information stored there. Because the SSH
aliases are likely to be configured only once by the script, then
skipped if the script is rerun on the Admin Workstation, add verbose
messages that state whether the task is being run or skipped. That
information will be valuable to an Admin trying to debug the Tails
environment configuration.