Updates Admin setup docs for new script config
Reformatting the guiding language in the post-install setup guide for
the Admin Workstation to advise Admins to reuse the A/THS files:

  * app-document-aths
  * app-source-ths

by copying them to the Transfer Device for reuse on the Journalist
Workstation. The `install.sh` script now supports reading those files
on either Workstation and configuring torrc accordingly.