Simplify docs for deleting old submissions
The goal of deleting old submissions is to minimize the amount of data
being backed up, thereby speeding up the tranfers over Tor and making
them less likely to fail. The passage was several paragraphs in length,
which is far too long for a "Note", so let's inline the text and trim
away excess verbiage. Admins will want facts.