Creates note about removing old backup formats
Succinct, just a gentle reminder that the old format archives can be
removed. The old archives were encrypted to the SecureDrop Application
GPG Key, which is of questionable security benefit. Additionally, the
old playbooks ran the backup task on every single playbook run, so it's
possible that long-running instances have a cluttered Admin Workstation
as a result. If Journalists weren't diligent about cleaning out
submissions, that could be quite a lot of data on a rather small USB
stick.