Skips backup role by default
Updates the prod and staging playbooks to skip the "backup" role. Since
it runs automatically every time the playbook is run, regardless of how
recently the last backup was created, it can rapidly take a substantial
amount of disk space, and delay successful completion of the playbook
due to the file transfers over Tor.

Let's instead skip it, allowing it to run under one of two conditions:

  * `perform_backup` var is set to true
  * `restore_file` is defined, meaning we'll want to restore

Updates the Backup and Restore topic guide in the docs to include
the new `ansible-playbook` invocation.