Updates git command from 'pull' -> 'fetch'
As of 0.3.7 we're requiring Tails v2.x. Tails v2, based on Debian
Jessie, ships with git v2, which changes the default strategy for pulls.
Hopefully most Admins will be proficient enough with git to navigate
the pull/fetch discrepancy, but let's err on the side of caution and
make the documentation as explicit as possible.

Using `git fetch` is terser and more legible than `git pull origin
release/0.3.7`, dig?