Mentions 3 USB port requirement in Tails upgrade docs
The docs assume that the SVS/airgapped machine has 3 USB ports available
for simultaneous use. All 3 USB devices must be plugged in together:

  * master Tails USB (running the OS)
  * target USB for upgrade
  * Backup Device (for storing persistence files)

Technically an Admin with only 2 USB ports available could use the
ephemeral storage on the Admin Workstation as a go-between for the
backup process, but that would require even more convoluted language in
the documentation.