Silences zenity output during script run
If the script needs to prompt for config information interactively,
`zenity` is called to display a dialog box and accept text input.
When called, zenity will print noisy output to STDERR:

  GtkDialog mapped without a transient parent. This is discouraged.
  ** (zenity:28513): WARNING **: atk-bridge: Couldn't listen on dbus
                     server: Failed to bind socket
                     "/root/.cache/at-spi2-socket-28513": No such file or
                     directory

Given that the install.sh script is designed to be run interactively in
a terminal, we should hide that expected error output so it doesn't
clutter up the otherwise informative output of the script. This has been
done by redirecting STDERR to /dev/null. Note that only STDERR is
redirected, since STDOUT must operate normally in order for the captured
text to percolate up to the function caller and be used in the script.