Finalizes Tails upgrade docs for 0.3.7 (#1317)
* Updates git command from 'pull' -> 'fetch'

As of 0.3.7 we're requiring Tails v2.x. Tails v2, based on Debian
Jessie, ships with git v2, which changes the default strategy for pulls.
Hopefully most Admins will be proficient enough with git to navigate
the pull/fetch discrepancy, but let's err on the side of caution and
make the documentation as explicit as possible.

Using `git fetch` is terser and more legible than `git pull origin
release/0.3.7`, dig?

* Mentions 3 USB port requirement in Tails upgrade docs

The docs assume that the SVS/airgapped machine has 3 USB ports available
for simultaneous use. All 3 USB devices must be plugged in together:

  * master Tails USB (running the OS)
  * target USB for upgrade
  * Backup Device (for storing persistence files)

Technically an Admin with only 2 USB ports available could use the
ephemeral storage on the Admin Workstation as a go-between for the
backup process, but that would require even more convoluted language in
the documentation.