Finalizes install.sh tails script (#1318)
* Silences zenity output during script run

If the script needs to prompt for config information interactively,
`zenity` is called to display a dialog box and accept text input.
When called, zenity will print noisy output to STDERR:

  GtkDialog mapped without a transient parent. This is discouraged.
  ** (zenity:28513): WARNING **: atk-bridge: Couldn't listen on dbus
                     server: Failed to bind socket
                     "/root/.cache/at-spi2-socket-28513": No such file or
                     directory

Given that the install.sh script is designed to be run interactively in
a terminal, we should hide that expected error output so it doesn't
clutter up the otherwise informative output of the script. This has been
done by redirecting STDERR to /dev/null. Note that only STDERR is
redirected, since STDOUT must operate normally in order for the captured
text to percolate up to the function caller and be used in the script.

* Removes verbose logging messages from install.sh

Partially reverts 69a7dacfd5158f8e4cc838254610d9d718e0c360. Since that's
a squashed commit, it's actually
8c63352f0327f9513cd4de8d133dc486594e2762 that's been reverted.

The `echo` statements were implemented during review and were
inadequately tested. The goal was to provide Admins with visibility into
where the script was referencing the A/THS URL information, since it
checks multiple locations. Unfortunately the `echo` statements caused
improper interpolation, and were also improperly called several times,
generating redundant output.

So let's remove the "Found here:" echo statements. Preserving the
echo calls that state that interactive prompting was necessary, although
those calls are now redirecting to STDERR.