Ignore strict version checking in dch
For this release's QA testing, I cut a release candidate with the
version "0.3.7+rc1". According to Debian's version numbering rules, this
is > "0.3.7", and dch prevents me from updating the changelog with a
version number that is technically later than the latest one already in
there.

0.3.7+rc1 was never pushed to production servers, so this is not going
to cause a problem in production. Testers who may have installed
0.3.7+rc1 will not automatically upgrade to 0.3.7, but that's only FPF
staff so it shouldn't matter.

In the future, we should use a versioning scheme that is compatible with
Debian's - I did not realize this was so complex. For now, I am
disabling the strict version checking in dch so I can get this release
out.