Fixes App play in prod playbook (develop) (#1341)
* Properly indents var include in prod playbook

Adds whitespace for consistent indentation level. Formatting only, no
functional change.

* Fixes App Server play in production playbook

The missing '-' was causing the play block for configuring the
Application Server, including installing the securedrop-app-code deb
package, to be skipped. When parsing the playbook, Ansible read the
subsequent play as part of the same YAML dict, which technically it was,
due to the missing hyphen. All app-code-related tests failed in the prod
environment as a result.