Re-include pycrypto dependency
When switching to our new pip-tools based Python dependency management system, I
managed to miss copying over the pycrypto line from our old standard
requirements file to our new .in "top-level" requirements file. This was not
caught in testing because the Ubuntu 14.04 Vagrant image we use for development
happens to include pycrypto, whereas the disk image of Ubuntu 14.04 does not.