Use catch-all regex for version numbers
The regexes in update_version.sh are sufficiently constrained by the content of
the surrounding lines, and it is not necessary to use a highly speific regex for
the actual format of the version numbers. Doing so in the past has created
issues when we experimented with new versioning schemas (e.g. 0.3.8-rc1). To
avoid this now and in the future, we change the version regexes in
update_version.sh to be generic.