Remove superfluous subparsers attribute settings
In my tests a singular subcommand was required even if the subparsers.required
line was not set.

As far as `subparsers.dest` goes, the function name is already saved to the
`parser.func` attribute which we call with `args.func()`. `subparsers.dest`
saves the title argument of the relevant `add_parser` subcommand to
`parser.subcommand` attribute, which is unused and thus unecessary.