UTF-8 encode plaintexts, fix unicode reply test
Since GPG operates on raw bytes, plaintext encryption input must be converted to
a stream/file-like type before it can be processed, and before that encoded as a
`str` if the plaintext is a `unicode`. `python-gnupg` handles this with a
function `gnupg._util._make_binary_stream()`. `python-gnupg`'s method of
choosing the encoding is based on locale settings, which we have observed to be
problematic in conjunction with how Apache creates Apache-controlled
environments for WSGI applications, such as SD (see #1343 and #1360--duplicate
issues). Explicitly calling it `_make_binary_stream` with "utf_8" on our
plaintexts before passing them to `gnupg.GPG.encrypt()` ensures encoding of
unicode runes will be handled correctly.

`python-gnupg` also calls `_make_binary_stream()` on ciphertext inputs to
`gnupg.GPG.decrypt()` and batch file inputs `gnupg.GPG.gen_key()` (without even
testing that the input's type belongs to `gnupg._util._STREAMLIKE_TYPES` with
`gnupg._util._is_stream()`). If we were to call `_make_binary_stream` on either
of these inputs before passing them to GPG, `python-gnupg` would try to
encode them, but as the input objects would now already be `_io.BytesIO`
streams, this would throw an exception.

For now it's safe to leave as is because ciphertexts are always returned as
`str`s from `python-gnupg` (or when read from encrypted files written by GPG)
and `gnupg.GPG.gen_key_input()`, which creates our batch file inputs, always
returns a `str` as well. Thus, we need not worry about encoding and
`python-gnupg` will handle conversion to a `_io.BytesIO` stream for us
internally.

It may be wise to try to make a pull request upstream to standardize the
behaviour of `python-gnupg`. It would be ideal to only pass `_io.BytesIO` stream
objects to `python-gnupg` in order to be as certain as possible we won't run
into further errors with `python-gnupg` processing our input `basestring`
objects, and instead our streams will be passed (more) directly to GPG.

The sad thing about this bug is that UTF-8 encoding had been explicitly enforced
on `unicode`s passed to our encryption function since the initial commit of
crypto_util.py in 62bbe59, and was just recently undone in 37ba11c. I believe I
this happened because I didn't realize Apache launched WSGI in a custom
environment (with LANG=C by default); from a reading of the `python-gnupg`
source code, and the fact our `test_unicode_reply` test passed, I was convinced
it was unecessary and removed it.

A related issue concerning a possible DoS against sources due to not explicitly
using UTF-8 for decoding after decryption (see
https://cure53.de/pentest-report_securedrop.pdf), introduced the
`tests.test_unit_integration.test_unicode_reply` test in 37f8f5e, which tries to
do pretty much what is says. The reason it failed to catch this regression is
because the test suite is designed to run in the development VM setting, which
does not use Apache. Thus, the LANG env var was, as is standard, set to UTF-8.
The test case has now been fixed to ensure that `python-gnupg` tries to use
`ansi_x3.4_1968` encoding.

Two more minor things:
* The object of decryption is now properly referred to as the ciphertext.
* Also, I changed the UTF-8 test string because look how awesome that
  Anglo-Saxon rune poem looks, right?